[rp] unit tests
