+++ src/modm/board/rp_pico/module.lb
@@ -89,14 +89,19 @@ def prepare(module, options):
         ":platform:gpio",
         ":platform:usb",
         ":platform:clockgen")
+
+    module.add_option(
+            BooleanOption(name="with_logger", description="Enable logging on UART0",
+                          default=False, dependencies=lambda v: ":platform:uart:0" if v else None))
     return True
 
 def build(env):
     env.outbasepath = "modm/src/modm/board"
     env.substitutions = {
-        "with_logger": False,
-        "with_assert": env.has_module(":architecture:assert")
+        "with_logger": env["with_logger"],
+        "with_assert": env.has_module(":architecture:assert"),
+        "board" : "rp-pico"
     }
+    env.template("board.hpp.in", "board.hpp")
     env.template("../board.cpp.in", "board.cpp")
-    env.copy('.')
     env.collect(":build:openocd.source", "interface/picoprobe.cfg", "target/rp2040.cfg")
