+++ test/modm/platform/pio/module.lb
@@ -1,0 +1,33 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+#
+# Copyright (c) 2026, Andrey Kunitsyn
+#
+# This file is part of the modm project.
+#
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+from collections import defaultdict, OrderedDict
+
+
+def init(module):
+    module.name = ":test:platform:pio"
+    module.description = "Tests for RP PIO"
+
+def prepare(module, options):
+    target = options[":target"]
+    if target.identifier.platform != "rp":
+        return False
+
+    module.depends(":platform:pio")
+    return True
+
+def build(env):
+   
+    device = env[":target"]
+    identifier = device.identifier
+   
+    env.outbasepath = "modm-test/src/modm-test/platform/pio"
+    env.copy("platform_pio_test_rp.cpp")
